Исправлены ошибки

+++ CarKits/Page1.xaml.cs
@@ -39,7 +39,7 @@ namespace CarKits
                     int maxspeed= int.Parse(sr.ReadLine());
                     string gearbox = sr.ReadLine();
                     string engine = sr.ReadLine();
-                    double weight = double.Parse(sr.ReadLine());
+                    int weight = int.Parse(sr.ReadLine());
                     chrc.Add(new Characteristics(brand, carcase, capacity, power, maxspeed, gearbox, engine, weight));
                 }
             foreach (Characteristics crist in chrc)
